+++ sdk/resourcemanager/azure-resourcemanager-resources/src/main/java/com/azure/resourcemanager/resources/implementation/FeatureImpl.java
@@ -7,6 +7,8 @@ import com.azure.resourcemanager.resources.fluentcore.model.implementation.Index
 import com.azure.resourcemanager.resources.models.Feature;
 import com.azure.resourcemanager.resources.fluent.models.FeatureResultInner;
 
+import java.util.regex.Pattern;
+
 /**
  * The implementation of {@link Feature}.
  */
@@ -35,6 +37,26 @@ final class FeatureImpl extends
             return null;
         }
         return innerModel().properties().state();
+    }
+
+    @Override
+    public String resourceProviderName() {
+        if (this.name() == null) {
+            return null;
+        }
+
+        String[] segments = this.name().split(Pattern.quote("/"));
+        return segments.length > 0 ? segments[0] : null;
+    }
+
+    @Override
+    public String featureName() {
+        if (this.name() == null) {
+            return null;
+        }
+
+        String[] segments = this.name().split(Pattern.quote("/"));
+        return segments.length > 1 ? segments[1] : null;
     }
 
     @Override