+++ sdk/resourcemanager/azure-resourcemanager-resources/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## 2.7.0-beta.1 (Unreleased)
 
+### Dependency Updates
+
+- Updated `api-version` of features to `2021-07-01`.
+
+### Features Added
+
+- Added support for `resourceProviderName` and `featureName` in `Feature`.
 
 ## 2.6.0 (2021-06-18)